+++ src/EventHub/EventHub/Cmdlets/AzureEventHubsCmdletBase.cs
@@ -106,6 +106,12 @@ namespace Microsoft.Azure.Commands.EventHub.Commands
         protected const string AliasResourceId = "ResourceId";
         protected const string AliasVirtualNetworkRule = "VirtualNteworkRule";
 
+        //Access Rights 
+        protected const string Manage = "Manage";
+        protected const string Send = "Send";
+        protected const string Listen = "Listen";
+
+
         protected struct SKU
         {
             internal const string Basic = "Basic";